arm64: shrink operand storage metadata
Store validated operand indexes in 32 bits and pack GC/EH root state with bounded value facts behind preserving accessors. This reduces storage from 32 to 24 bytes and elem from 72 to 64 bytes without changing generated code.\n\nNative json-as compilation drops 33,634 B/op (11.6%) and many_funcs drops 640 B/op with allocation counts unchanged. The full native ARM64 suite passes and representative code hashes are identical.

+++ src/core/compiler/backend/railshot/arm64/interval_region.go
@@ -146,7 +146,7 @@ func (f *fn) intervalLocalHasMemBorrow(x int) bool {
 
 func (f *fn) demoteIntervalLocalRefs(x int) {
 	for e := f.s.head.next; e != f.s.head; e = e.next {
-		if e.kind == ekValue && e.st.kind == stLocalReg && e.st.idx == x {
+		if e.kind == ekValue && e.st.kind == stLocalReg && e.st.idx == uint32(x) {
 			e.st.kind = stLocalRef
 			e.st.reg = regNone
 		}
